+++ skorch/tests/test_classifier.py
@@ -289,7 +289,12 @@ class TestNeuralNetBinaryClassifier:
         mock = Mock(side_effect=lambda x: x)
         monkeypatch.setattr(torch, "sigmoid", mock)
 
-        net = net_cls(module_cls, max_epochs=1, lr=0.1, criterion=nn.MSELoss)
+        # we need to add a custom nonlinearity, otherwise the output won't be 2d
+        def nonlin(x):
+            return torch.stack((1 - x, x), axis=1)
+
+        net = net_cls(module_cls, max_epochs=1, lr=0.1, criterion=nn.MSELoss,
+                      predict_nonlinearity=nonlin)
         X, y = data
         net.fit(X, y)